pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.53.0",
    "langchain>=1.1.2",
    "langchain-openai>=1.1.0",
    "pyyaml>=6.0.3",
]
